refactor: move each handler into the file that owns its feature
