+++ main.go
@@ -219,46 +219,6 @@ func pagination(c *fiber.Ctx) (page, limit int, ok bool) {
 	return page, limit, true
 }
 
-// ftsSyntaxErrors are the messages SQLite returns when the *caller's* search
-// string is not a valid FTS5 expression, as opposed to when something is wrong
-// with the server. Measured against the real artifact:
-//
-//	AND | ) | NOT | ()      fts5: syntax error near "…"
-//	*                       unknown special query:
-//	" (unbalanced)          unterminated string
-//
-// The SQL around the MATCH is a fixed string, so the only variable in these
-// statements is what the caller sent. Matching on the message is unpleasant —
-// the driver offers no error code for this — but the alternative is answering
-// 500 to a typo, which makes every 5xx alert untrustworthy.
-var ftsSyntaxErrors = []string{
-	"fts5: syntax error",
-	"unknown special query",
-	"unterminated string",
-	"malformed MATCH expression",
-}
-
-func isFTSSyntaxError(err error) bool {
-	msg := strings.ToLower(err.Error())
-	for _, s := range ftsSyntaxErrors {
-		if strings.Contains(msg, s) {
-			return true
-		}
-	}
-	return false
-}
-
-// badSearchQuery answers a malformed search expression. It is logged at warn,
-// not error: the server is fine, the query was not.
-func badSearchQuery(c *fiber.Ctx, query, lang string, err error) error {
-	zlog.Warn().Str("query", query).Str("lang", lang).Err(err).Msg("ftsSearchHandler: malformed search query")
-	return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{
-		"error": "Invalid search query: it is not a valid full-text search expression. " +
-			"Bare operators (AND, OR, NOT), unbalanced parentheses or quotes, and a " +
-			"leading * are all rejected by SQLite. Quote the text to search for it literally.",
-	})
-}
-
 // queryFailed renders a failed query. A deadline is returned as an error so the
 // timeout middleware can answer 408, which is worth distinguishing from the 500
 // that a genuinely broken query earns.
@@ -301,502 +261,6 @@ func authMiddleware(c *fiber.Ctx) error {
 	}
 
 	return c.Status(fiber.StatusUnauthorized).JSON(fiber.Map{"error": "Invalid API key"})
-}
-
-// Handler for /api/events/:id
-func getEventHandler(c *fiber.Ctx) error {
-	lang := c.Query("lang", "en")
-	db := dbFor(c)
-	id := c.Params("id")
-
-	zlog.Info().Str("id", id).Str("lang", lang).Msg("getEventHandler called")
-
-	eventID, err := strconv.ParseUint(id, 10, 32)
-	if err != nil {
-		zlog.Warn().Str("id", id).Str("lang", lang).Err(err).Msg("getEventHandler: Invalid Event ID format")
-		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{
-			"error": "Invalid Event ID format",
-		})
-	}
-
-	var event Event
-	result := db.First(&event, uint(eventID))
-
-	if result.Error != nil {
-		if errors.Is(result.Error, gorm.ErrRecordNotFound) {
-			zlog.Warn().Str("id", id).Str("lang", lang).Err(result.Error).Msg("getEventHandler: Event not found")
-			return c.Status(fiber.StatusNotFound).JSON(fiber.Map{
-				"error": "Event not found",
-			})
-		}
-		zlog.Error().Str("id", id).Str("lang", lang).Err(result.Error).Msg("getEventHandler: Failed to retrieve event")
-		return queryFailed(c, result.Error, "Failed to retrieve event")
-	}
-	zlog.Info().Str("id", id).Str("lang", lang).Msg("getEventHandler: Successfully retrieved event")
-	return c.JSON(fiber.Map{"data": event})
-}
-
-// Structure for the /api/tags response
-type TagInfo struct {
-	Tag   string `json:"tag"`
-	Count int    `json:"count"`
-}
-
-// Structure for the /api/categories response
-type CategoryInfo struct {
-	Category string `json:"category"`
-	Count    int    `json:"count"`
-}
-
-// Handler for /api/categories — the counterpart to /api/tags, and what a client
-// needs to build a filter UI without fetching every event to discover what
-// exists.
-//
-// Simpler than getTagsHandler in one way that matters: tags are a JSON array,
-// so that query joins through json_each and must COUNT(DISTINCT e.id) to avoid
-// counting a row twice when it lists a tag twice. category is exactly one value
-// per row, so a plain COUNT(*) is already an event count and cannot disagree
-// with /api/events?category=.
-func getCategoriesHandler(c *fiber.Ctx) error {
-	lang := c.Query("lang", "en")
-	db := dbFor(c)
-
-	zlog.Info().Str("lang", lang).Msg("getCategoriesHandler called")
-
-	// An artifact predating the category column has nothing to report, and the
-	// statement below would fail against it with `no such column`. Answering the
-	// empty list is the truth — this artifact carries no categories — and a 500
-	// on an endpoint the service can perfectly well answer is not.
-	if !categoriesByLang[resolveLang(lang)].present {
-		return c.JSON(fiber.Map{"data": []CategoryInfo{}})
-	}
-
-	// Initialised, not declared nil, for the reason ftsSearchHandler spells out:
-	// Raw().Scan() leaves the slice nil when nothing matches and a nil slice
-	// marshals to JSON null, so a caller would get null from one branch of this
-	// handler and [] from the other.
-	result := []CategoryInfo{}
-	// NOTHING MAY FOLLOW THE FINAL SEMICOLON — see getTagsHandler for what a
-	// trailing comment does to this driver. It is not a style rule.
-	sqlQuery := `
-SELECT
-    LOWER(TRIM(category)) AS category,
-    COUNT(*) AS count
-FROM
-    events
-WHERE
-    category IS NOT NULL
-    AND TRIM(category) != ''
-GROUP BY
-    LOWER(TRIM(category))
-ORDER BY
-    category ASC;`
-	if err := db.Raw(sqlQuery).Scan(&result).Error; err != nil {
-		zlog.Error().Str("lang", lang).Err(err).Msg("getCategoriesHandler: Error executing raw SQL for categories")
-		return queryFailed(c, err, "Failed to retrieve categories from database")
-	}
-
-	zlog.Info().Int("category_count", len(result)).Str("lang", lang).Msg("getCategoriesHandler: Successfully retrieved categories")
-	return c.JSON(fiber.Map{"data": result})
-}
-
-// Handler for /api/tags
-func getTagsHandler(c *fiber.Ctx) error {
-	lang := c.Query("lang", "en")
-	db := dbFor(c)
-
-	zlog.Info().Str("lang", lang).Msg("getTagsHandler called")
-
-	// Initialised, not declared nil, for the reason spelled out in
-	// ftsSearchHandler: Raw().Scan() leaves the slice nil when nothing matches
-	// and a nil slice marshals to JSON null, so a caller would have to
-	// special-case one endpoint. /api/categories was written this way from the
-	// start; this one is its sibling and must not answer in a different shape.
-	// Reachable only on an artifact where no row carries a usable tag.
-	result := []TagInfo{}
-	// NOTHING MAY FOLLOW THE FINAL SEMICOLON — not even a comment. SQLite
-	// prepares the text after it as a further statement, and a comment-only
-	// statement prepares successfully to a NULL handle. go-sqlite3 then steps
-	// that NULL handle, gets neither DONE nor ROW, and returns nil instead of
-	// io.EOF (sqlite3.go:2238), so database/sql calls Next forever. The
-	// handler hangs, holding the connection, with no error and no timeout.
-	// This endpoint hung for exactly that reason. Comments anywhere before the
-	// final semicolon are fine.
-	sqlQuery := `
-SELECT
-    LOWER(j.value) AS tag,
-    -- DISTINCT because this endpoint counts events, not occurrences. Four rows
-    -- once listed the same tag twice (RU 160/231 'price', RU 333 and EN 155
-    -- 'bitcoin'), so COUNT(*) read 446 for 'bitcoin' against the 445 events
-    -- that carried it. Canonical normalised those rows and now rejects
-    -- duplicates in validator invariant 6, so the discrepancy should not recur
-    -- — but the API is not the right place to depend on that.
-    COUNT(DISTINCT e.id) AS count
-FROM
-    events e,
-    json_each(e.tags) j
-WHERE
-    e.tags IS NOT NULL
-    AND e.tags != ''        -- Not an empty string literal
-    AND e.tags != '[]'      -- Not an empty JSON array string literal
-    AND json_valid(e.tags) = 1 -- Ensures the string is valid JSON
-    AND json_type(e.tags) = 'array' -- Ensures it's specifically a JSON array
-    AND j.value IS NOT NULL
-    AND TRIM(CAST(j.value AS TEXT)) != '' -- Ensures the extracted tag is not an empty or whitespace-only string
-GROUP BY
-    LOWER(j.value) -- Group by the lowercased tag for case-insensitive counting
--- Order alphabetically by the (now lowercased) tag
-ORDER BY
-    tag ASC;`
-	if err := db.Raw(sqlQuery).Scan(&result).Error; err != nil {
-		zlog.Error().Str("lang", lang).Err(err).Msg("getTagsHandler: Error executing raw SQL for tags")
-		return queryFailed(c, err, "Failed to retrieve tags from database")
-	}
-
-	zlog.Info().Int("tag_count", len(result)).Str("lang", lang).Msg("getTagsHandler: Successfully retrieved tags")
-	return c.JSON(fiber.Map{"data": result})
-}
-
-// Handler for /api/events/tags/:tag
-func getEventsByTagHandler(c *fiber.Ctx) error {
-	lang := c.Query("lang", "en")
-	db := dbFor(c)
-	tagParam := c.Params("tag")
-
-	zlog.Info().Str("tag", tagParam).Str("lang", lang).Msg("getEventsByTagHandler called")
-
-	// The tag filter does not compose with the category or landmark filters.
-	// See filterParamRejected.
-	if filterParamRejected(c, "category", "landmark") {
-		return nil
-	}
-
-	page, limit, ok := pagination(c)
-	if !ok {
-		return nil
-	}
-	offset := (page - 1) * limit
-
-	var events []Event
-	var totalEvents int64
-
-	// Match the tag as a JSON array element, the same way /api/tags counts
-	// them, so the two endpoints cannot disagree about what a tag is.
-	//
-	// This replaces a LIKE '%"tag"%' substring match against the raw JSON. That
-	// matched the same rows for every tag in the current vocabulary, but it
-	// interpreted LIKE metacharacters in the caller's input: /api/events/tags/%
-	// returned all 582 RU events and /api/events/tags/_____ returned 246,
-	// because % and _ are wildcards. An equality test inside json_each cannot
-	// be steered that way, and it stops depending on how the JSON is spaced or
-	// quoted.
-	tagMatch := `EXISTS (
-		SELECT 1 FROM json_each(events.tags) j
-		WHERE json_valid(events.tags) = 1
-		  AND json_type(events.tags) = 'array'
-		  AND LOWER(j.value) = ?
-	)`
-	searchTag := strings.ToLower(tagParam)
-
-	countQuery := db.Model(&Event{}).Where(tagMatch, searchTag)
-	if err := countQuery.Count(&totalEvents).Error; err != nil {
-		zlog.Error().Str("tag", tagParam).Str("lang", lang).Err(err).Msg("getEventsByTagHandler: Failed to count events by tag")
-		return queryFailed(c, err, "Failed to count events by tag")
-	}
-
-	// Get paginated events matching the tag, newest first. See eventOrder for
-	// why the sort does not stop at the date.
-	dataQuery := db.Model(&Event{}).Order(eventOrder).Limit(limit).Offset(offset).Where(tagMatch, searchTag)
-	if err := dataQuery.Find(&events).Error; err != nil {
-		zlog.Error().Str("tag", tagParam).Str("lang", lang).Int("page", page).Int("limit", limit).Err(err).Msg("getEventsByTagHandler: Failed to retrieve events by tag")
-		return queryFailed(c, err, "Failed to retrieve events by tag")
-	}
-
-	totalPages := (totalEvents + int64(limit) - 1) / int64(limit)
-	zlog.Info().Int("event_count", len(events)).Str("tag", tagParam).Str("lang", lang).Int("page", page).Int("limit", limit).Int64("total_matching", totalEvents).Msg("getEventsByTagHandler: Successfully retrieved events")
-
-	return c.JSON(PaginatedEventsResponse{
-		Events: events,
-		Pagination: PaginationData{
-			CurrentPage: page,
-			LastPage:    int(totalPages),
-			PerPage:     limit,
-			Total:       totalEvents,
-		},
-	})
-}
-
-// Handler for /api/events
-func getAllEventsHandler(c *fiber.Ctx) error {
-	lang := c.Query("lang", "en")
-	db := dbFor(c)
-	yearStr := c.Query("year")
-	monthStr := c.Query("month")
-	dayStr := c.Query("day")
-
-	zlog.Info().Str("lang", lang).Str("year", yearStr).Str("month", monthStr).Str("day", dayStr).Msg("getAllEventsHandler called")
-
-	page, limit, ok := pagination(c)
-	if !ok {
-		return nil
-	}
-	offset := (page - 1) * limit
-
-	events := []Event{}
-	var totalEvents int64
-	query := db.Model(&Event{})
-
-	// Date filters are validated rather than passed through, because an
-	// unparseable value here matches nothing and the caller gets 200 with an
-	// empty list — identical to a day that genuinely has no events. A bot on
-	// that response posts nothing and reports success, so a typo in its date
-	// arithmetic would look exactly like a quiet day, indefinitely.
-	if yearStr != "" {
-		if !isNumericInRange(yearStr, 1000, 9999) {
-			return badParam(c, "year", yearStr, "a four-digit year")
-		}
-		query = query.Where("strftime('%Y', date) = ?", yearStr)
-	}
-	if monthStr != "" {
-		if !isNumericInRange(monthStr, 1, 12) {
-			return badParam(c, "month", monthStr, "1–12")
-		}
-		// Two digits, to match what strftime('%m') returns. Both "1" and "01"
-		// are accepted from the caller.
-		query = query.Where("strftime('%m', date) = ?", pad2(monthStr))
-	}
-	if dayStr != "" {
-		if !isNumericInRange(dayStr, 1, 31) {
-			return badParam(c, "day", dayStr, "1–31")
-		}
-		query = query.Where("strftime('%d', date) = ?", pad2(dayStr))
-	}
-
-	// category is validated against the vocabulary this artifact actually
-	// carries, read at boot by loadCategories. An unknown value is a 400 for
-	// the same reason a malformed month is: answering 200 with an empty list
-	// makes "there is no such category" indistinguishable from "that category
-	// has no events", and a client cannot tell a typo from a quiet corner of
-	// the corpus.
-	//
-	// Matched case-insensitively, like the tag filter, and lowercased on the
-	// way in because every stored value is lowercase.
-	if categoryStr := c.Query("category"); categoryStr != "" {
-		want := strings.ToLower(strings.TrimSpace(categoryStr))
-		// resolveLang, not the raw parameter: the vocabulary consulted must be
-		// the one belonging to the artifact this request will actually read.
-		vocab := categoriesByLang[resolveLang(lang)]
-		if !vocab.known(want) {
-			return badParam(c, "category", categoryStr, vocab.expected())
-		}
-		// LOWER on the column, not a bare equality: the closed set is enforced
-		// by the publisher rather than by the schema, so this must not depend
-		// on the stored casing being what it is today.
-		query = query.Where("LOWER(TRIM(category)) = ?", want)
-	}
-
-	// landmark is the switch the website calls «Только главное»: one boolean,
-	// orthogonal to category, hiding everything that is not a landmark. It ANDs
-	// with category and with the date filters, so ?category=tech&landmark=true
-	// is "the tech events that matter".
-	//
-	// Presence is checked before the value is parsed, and that order is the
-	// point. Against an artifact predating the column — a rollback target —
-	// `WHERE landmark = ?` is `no such column: landmark`, measured, which would
-	// be a 500 for a question this service can answer perfectly well. It is also
-	// the more useful of the two rejections: "this artifact predates the column"
-	// tells the caller something "expected true or false" cannot.
-	//
-	// Unlike category there is no vocabulary to consult, so an unparseable value
-	// is the only other way to be wrong. It is a 400 rather than a quiet default
-	// for the reason the date filters are: ?landmark=yes silently read as false
-	// answers 200 with the 179 rows the caller least wanted, and nothing in the
-	// response says the filter was not the one asked for.
-	if landmarkStr := c.Query("landmark"); landmarkStr != "" {
-		// resolveLang, not the raw parameter, for the reason the category filter
-		// spells out: the artifact consulted must be the one this request reads.
-		flag := landmarkByLang[resolveLang(lang)]
-		if !flag.present {
-			return badParam(c, "landmark", landmarkStr, flag.expected())
-		}
-		// ParseBool rather than a hand-rolled comparison: it is the spelling a Go
-		// client would produce and a documented set (1/t/T/TRUE/true/True and the
-		// false equivalents), so callers who send "1" are not surprised. The
-		// rejection names `true or false` because that is the canonical form to
-		// reach for, not because the others are refused.
-		want, err := strconv.ParseBool(strings.TrimSpace(landmarkStr))
-		if err != nil {
-			return badParam(c, "landmark", landmarkStr, flag.expected())
-		}
-		// No LOWER/TRIM counterpart here: this column is INTEGER NOT NULL, so
-		// unlike category there is no stored casing or padding to defend
-		// against. Bound as a Go bool, which the driver binds as 1 or 0.
-		query = query.Where("landmark = ?", want)
-	}
-
-	if err := query.Count(&totalEvents).Error; err != nil {
-		zlog.Error().Str("lang", lang).Err(err).Msg("getAllEventsHandler: Failed to count events")
-		return queryFailed(c, err, "Failed to count events")
-	}
-
-	if err := query.Order(eventOrder).Limit(limit).Offset(offset).Find(&events).Error; err != nil {
-		zlog.Error().Str("lang", lang).Err(err).Msg("getAllEventsHandler: Failed to retrieve events")
-		return queryFailed(c, err, "Failed to retrieve events")
-	}
-
-	totalPages := (totalEvents + int64(limit) - 1) / int64(limit)
-
-	zlog.Info().Int("event_count", len(events)).Int64("total_matching", totalEvents).Str("lang", lang).Msg("getAllEventsHandler: Successfully retrieved events")
-
-	return c.JSON(PaginatedEventsResponse{
-		Events: events,
-		Pagination: PaginationData{
-			CurrentPage: page,
-			LastPage:    int(totalPages),
-			PerPage:     limit,
-			Total:       totalEvents,
-		},
-	})
-}
-
-// Handler for /api/search
-func ftsSearchHandler(c *fiber.Ctx) error {
-	lang := c.Query("lang", "en")
-	db := dbFor(c)
-	query := c.Query("q")
-
-	zlog.Info().Str("query", query).Str("lang", lang).Msg("ftsSearchHandler called")
-
-	if query == "" {
-		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{"error": "Search query is required"})
-	}
-
-	// Search does not narrow by category or landmark. See filterParamRejected.
-	if filterParamRejected(c, "category", "landmark") {
-		return nil
-	}
-
-	page, limit, ok := pagination(c)
-	if !ok {
-		return nil
-	}
-	offset := (page - 1) * limit
-
-	// Initialised, not declared nil: this handler scans into the slice with
-	// Raw().Scan(), which leaves it nil when nothing matches, and a nil slice
-	// marshals to JSON null. Every other endpoint uses Find(), which sets an
-	// empty slice, so search was the one place a caller got null instead of []
-	// and had to special-case it.
-	events := []Event{}
-	var totalEvents int64
-
-	// The caller's string is handed to FTS5 as written. It used to be passed
-	// through strings.ReplaceAll(query, `"`, `""`), which was not sanitisation —
-	// the value is already a bound parameter, so there is no injection to
-	// prevent, and MATCH takes an expression rather than a literal. What the
-	// doubling actually did was destroy phrase search: `"bitcoin price"` became
-	// `""bitcoin price""`, which FTS5 reads as an empty phrase followed by two
-	// bare tokens, i.e. an implicit AND. Against the production artifact that
-	// answered 39 where the phrase matches 6, and `"price bitcoin"` returned the
-	// same 39 rather than its own 23 — word order silently ignored, on the one
-	// syntax the documentation tells callers to reach for. It also re-balanced
-	// every stray quote, so `bitcoin"` answered 200 instead of the documented
-	// 400.
-	//
-	// Malformed expressions need no pre-filtering here: SQLite rejects them and
-	// isFTSSyntaxError turns that into a 400.
-	countSQL := `
-		SELECT COUNT(*)
-		FROM events e
-		JOIN events_fts fts ON e.id = fts.rowid
-		WHERE events_fts MATCH ?;
-	`
-	if err := db.Raw(countSQL, query).Scan(&totalEvents).Error; err != nil {
-		if isFTSSyntaxError(err) {
-			return badSearchQuery(c, query, lang, err)
-		}
-		zlog.Error().Str("query", query).Str("lang", lang).Err(err).Msg("ftsSearchHandler: Failed to count search results")
-		return queryFailed(c, err, "Failed to count search results")
-	}
-
-	// The two columns in the list below that an artifact may genuinely not have.
-	// A release predating 2026-08-09 has no `category` and one predating
-	// 2026-08-12 has no `landmark`, and those files are rollback targets — so
-	// naming either unconditionally makes every search on a rolled-back artifact
-	// answer 500, which is the outage the boot path was taught not to cause.
-	// Every other handler adapts for free: GORM builds its SELECT from the
-	// struct against the table it can see and leaves the field at its zero value
-	// — measured against an artifact with both columns dropped — and only this
-	// statement enumerates by hand.
-	//
-	// They are independent rather than a single "old artifact" flag, because the
-	// two dates are two months' worth of releases apart and an artifact between
-	// them has one column and not the other.
-	//
-	// Spliced rather than parameterised because a column name is not a bind
-	// parameter. The value is built here from literals, never from anything a
-	// caller sent. Sprintf is safe on this statement specifically because it
-	// contains no other %% verb; anything added below that needs one must escape
-	// it or this stops being a formatting string.
-	optionalCols := ""
-	if categoriesByLang[resolveLang(lang)].present {
-		optionalCols += " e.category,"
-	}
-	if landmarkByLang[resolveLang(lang)].present {
-		optionalCols += " e.landmark,"
-	}
-
-	searchSQL := fmt.Sprintf(`
-		-- "references" is a SQL reserved word: unquoted, SQLite refuses to
-		-- parse the statement and every search returns 500.
-		--
-		-- This list is the one place in the service that names columns by hand;
-		-- every other handler goes through db.Model(&Event{}) and picks up a new
-		-- field for free. So this is the statement that silently drops one, and
-		-- it had dropped three: category, created_at and updated_at all reached
-		-- callers empty while /api/events returned them correctly.
-		-- TestEventStructCoversEveryColumn covers search specifically for that
-		-- reason, and checks the values rather than the keys, because a struct
-		-- field the SELECT never fetched still marshals to "", null — or, since
-		-- landmark, to false, which is why that test had to learn that a false
-		-- proves nothing about whether the column was fetched.
-		--
-		-- fts.rank is deliberately not selected: nothing scans it — there is no
-		-- Rank field on Event — and SQLite orders by it perfectly well without
-		-- it appearing here. Selecting it only implied a field that does not
-		-- exist. Exposing rank as JSON would be a new contract decision: the
-		-- value is FTS5-internal, negative, and incomparable across queries.
-		SELECT e.id, e.date, e.title, e.description, e.tags, e.media, e."references",
-		       e.url_path,%s e.created_at, e.updated_at
-		FROM events e
-		JOIN events_fts fts ON e.id = fts.rowid
-		WHERE events_fts MATCH ?
-		-- e.id breaks ties in rank. Without it equally-ranked rows come back in
-		-- whatever order the scan produces, which is not required to be the same
-		-- order on the next request — so a caller walking pages can be handed one
-		-- event twice and never shown another.
-		ORDER BY fts.rank, e.id DESC
-		LIMIT ? OFFSET ?;
-	`, optionalCols)
-	if err := db.Raw(searchSQL, query, limit, offset).Scan(&events).Error; err != nil {
-		if isFTSSyntaxError(err) {
-			return badSearchQuery(c, query, lang, err)
-		}
-		zlog.Error().Str("query", query).Str("lang", lang).Err(err).Msg("ftsSearchHandler: Failed to execute search")
-		return queryFailed(c, err, "Failed to execute search")
-	}
-
-	totalPages := (totalEvents + int64(limit) - 1) / int64(limit)
-
-	return c.JSON(PaginatedEventsResponse{
-		Events: events,
-		Pagination: PaginationData{
-			CurrentPage: page,
-			LastPage:    int(totalPages),
-			PerPage:     limit,
-			Total:       totalEvents,
-		},
-	})
 }
 
 func getAllowedOrigins() string {
